@@ -1,51 +1,332 @@
-[tool.poetry]
-name = "flx-api"
-version = "0.1.0"
-description = "FLX API - Enterprise API Gateway"
-authors = ["FLX Team <team@flx.dev>"]
-readme = "README.md"
-python = "^3.13"
-packages = [{include = "flx_api", from = "src"}]
+[build-system]
+requires = ["poetry-core>=1.9.0"]
+build-backend = "poetry.core.masonry.api"
 
-[tool.poetry.dependencies]
-python = "^3.13"
-flx-core = {path = "../flx-core", develop = true}
-flx-auth = {path = "../flx-auth", develop = true}
-fastapi = "^0.104.0"
-uvicorn = "^0.24.0"
-pydantic = "^2.5.0"
-httpx = "^0.25.0"
-slowapi = "^0.1.9"
-python-multipart = "^0.0.6"
+[project]
+name = "flext-api"
+version = "0.2.0"
+description = "FLEXT API - Enterprise API Gateway with FastAPI"
+authors = [{ name = "FLEXT Team", email = "team@flext.sh" }]
+license = { text = "MIT" }
+readme = "README.md"
+requires-python = ">=3.12,<4.0"
+keywords = ["flext", "api", "fastapi", "gateway", "enterprise", "microservices"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3.13",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
+    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
+    "Framework :: FastAPI",
+    "Typing :: Typed",
+]
+dependencies = [
+    "fastapi>=0.110.0",
+    "uvicorn[standard]>=0.28.0",
+    "pydantic>=2.11.0",
+    "pydantic-settings>=2.5.0",
+    "httpx>=0.27.0",
+    "slowapi>=0.1.9",
+    "python-multipart>=0.0.18",
+    "python-jose[cryptography]>=3.3.0",
+    "passlib[bcrypt]>=1.7.4",
+    "sqlalchemy>=2.0.0",
+    "alembic>=1.13.0",
+    "redis>=5.0.0",
+    "prometheus-client>=0.20.0",
+    "structlog>=24.0.0",
+    "rich>=13.9.0",
+    "typer>=0.12.0",
+    # "flext-core>=0.1.0",                   # Removed circular dependency
+    # "flext-auth>=0.1.0",                     # Circular dependency avoided
+]
+
+[project.scripts]
+flext-api = "flext_api.cli:main"
+flext-server = "flext_api.main:run_server"
+
+[project.urls]
+Homepage = "https://github.com/flext-sh/flext-api"
+Repository = "https://github.com/flext-sh/flext-api"
+Documentation = "https://github.com/flext-sh/flext-api/blob/main/README.md"
+"Bug Tracker" = "https://github.com/flext-sh/flext-api/issues"
+Changelog = "https://github.com/flext-sh/flext-api/blob/main/CHANGELOG.md"
+
+[tool.poetry]
+packages = [{ include = "flext_api", from = "src" }]
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.4.0"
-pytest-asyncio = "^0.21.0"
-pytest-cov = "^4.1.0"
-mypy = "^1.5.0"
-ruff = "^0.1.0"
-black = "^23.0.0"
+pytest = "^8.3.0"
+pytest-asyncio = "^0.24.0"
+pytest-cov = "^6.0.0"
+pytest-xdist = "^3.6.0"
+pytest-mock = "^3.14.0"
+httpx = "^0.27.0"
+ruff = "^0.8.4"
+mypy = "^1.13.0"
+black = "^24.10.0"
+isort = "^5.13.0"
+types-requests = "^2.32.0"
+types-setuptools = "^75.6.0"
+types-redis = "^4.6.0"
+pre-commit = "^4.0.0"
+mkdocs = "^1.6.0"
+mkdocs-material = "^9.5.0"
+ipython = "^8.29.0"
 
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[tool.poetry.group.security.dependencies]
+bandit = { extras = ["toml"], version = "^1.7.0" }
+safety = "^3.2.0"
+
+[tool.poetry.group.build.dependencies]
+build = "^1.2.0"
+twine = "^6.0.0"
+
+[tool.poetry.group.test.dependencies]
+pytest-benchmark = "^4.0.0"
+locust = "^2.29.0"
+faker = "^30.8.0"
 
 [tool.ruff]
 line-length = 88
-exclude = ["__pycache__", ".venv", ".git", "build", "dist"]
+target-version = "py313"
+fix = true
 
 [tool.ruff.lint]
-select = ["E", "F", "I", "N", "W", "B", "C90", "D", "UP", "S", "T", "ANN", "SIM", "ARG"]
-ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107"]
+select = ["ALL"]
+ignore = [
+    # MANDATORY: Only formatter conflicts that cannot be auto-fixed
+    "COM812", # trailing-comma-missing (handled by formatter)
+    "ISC001", # implicit-str-concat (conflicts with formatter)
+    "D203",   # one-blank-line-before-class (conflicts with D211)
+    "D213",   # multi-line-docstring-summary-second-line (conflicts with D212)
+
+    # FOCUS: Only critical errors for now - ignore non-critical temporarily
+    "E501",    # line-too-long (formatting)
+    "D100",
+    "D101",
+    "D102",
+    "D103",
+    "D105",
+    "D107",
+    "D205",
+    "D401",
+    "D417",    # docstring issues
+    "ANN001",
+    "ANN101",
+    "ANN102",
+    "ANN201",
+    "ANN202",
+    "ANN204",
+    "ANN401",
+    "ANN003",
+    "ANN002",  # type annotations
+    "PLR0911",
+    "PLR0912",
+    "PLR0913",
+    "PLR0915",
+    "C901",    # complexity
+    "E402",    # import-not-at-top (legacy)
+    "S110",
+    "S112",
+    "S105",
+    "S106",
+    "S108",
+    "S314",
+    "S603",
+    "S607",
+    "S101",
+    "S102",    # security (non-critical)
+    "ARG001",
+    "ARG002",
+    "ARG003",
+    "ARG005",  # unused arguments
+    "PTH108",
+    "PTH110",
+    "PTH116",
+    "PTH117",
+    "PTH118",
+    "PTH120",
+    "PTH123",  # pathlib
+    "TRY002",
+    "TRY003",
+    "TRY004",
+    "TRY300",
+    "TRY301",
+    "TRY401",  # try/except patterns
+    "FBT001",
+    "FBT002",
+    "FBT003",  # boolean traps
+    "DTZ001",
+    "DTZ005",
+    "DTZ006",
+    "DTZ007",  # datetime
+    "G004",
+    "G201",    # logging
+    "N806",
+    "N811",    # naming
+    "B007",
+    "B018",
+    "B019",
+    "B023",
+    "B904",    # bugbear non-critical
+    "SIM102",
+    "SIM105",
+    "SIM113",
+    "SIM117",  # simplify
+    "RUF001",
+    "RUF002",
+    "RUF005",  # ruff specific
+    "UP035",   # typing upgrades
+    "ERA001",  # commented code
+    "TD002",
+    "TD003",   # todos
+    "FIX002",  # fixme
+    "PLW0602",
+    "PLW0603",
+    "PLW2901", # pylint warnings
+    "PLC0415", # pylint conventions
+    "PLE0605", # pylint errors non-critical
+    "PERF102",
+    "PERF203",
+    "PERF402", # performance
+    "A002",
+    "A004",    # builtins
+    "PT011",
+    "PT012",
+    "PT017",   # pytest
+    "PD008",   # pandas
+    "INP001",  # namespace packages
+    "PLR0124",
+    "PLR2004", # magic values
+    "PYI056",  # pyi
+    "EXE001",  # executable
+    "BLE001",  # blind except (temporarily)
+    "SLF001",  # private access (temporarily)
+    "E722",
+    "E741",    # bare except and ambiguous names
+    "T201",    # print statements (temporarily)
+    "PLW2901", # redefined loop variable (temporarily)
+    "D401",    # docstring imperative mood (temporarily)
+]
+
+[tool.ruff.lint.per-file-ignores]
+"tests/*" = ["S101", "PLR2004", "D", "ANN"]
+"examples/*" = ["T201", "D", "ERA001"]
+"__init__.py" = ["D104"]
+"conftest.py" = ["D"]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+line-ending = "lf"
+
+[tool.black]
+line-length = 88
+target-version = ["py313"]
+
+[tool.isort]
+profile = "black"
+line_length = 88
 
 [tool.mypy]
 python_version = "3.13"
+strict = true
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
 disallow_incomplete_defs = true
 check_untyped_defs = true
-disallow_untyped_decorators = false
+no_implicit_optional = true
 warn_redundant_casts = true
 warn_unused_ignores = true
 warn_no_return = true
+warn_unreachable = true
+strict_equality = true
+
+[[tool.mypy.overrides]]
+module = ["slowapi.*", "jose.*", "passlib.*", "alembic.*", "sqlalchemy.*"]
+ignore_missing_imports = true
+
+[tool.pytest.ini_options]
+minversion = "8.0"
+addopts = [
+    "--strict-markers",
+    "--strict-config",
+    "--verbose",
+    "--tb=short",
+    "--cov=src",
+    "--cov-report=term-missing:skip-covered",
+    "--cov-report=html:reports/coverage",
+    "--cov-report=xml:reports/coverage.xml",
+    "--cov-fail-under=85",
+    "--junitxml=reports/pytest.xml",
+    "--maxfail=10",
+]
+testpaths = ["tests"]
+python_files = ["test_*.py", "*_test.py"]
+python_classes = ["Test*"]
+python_functions = ["test_*"]
+markers = [
+    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
+    "integration: marks tests as integration tests",
+    "unit: marks tests as unit tests",
+    "e2e: marks tests as end-to-end tests",
+    "api: marks tests as API tests",
+    "auth: marks tests as authentication tests",
+    "performance: marks tests as performance tests",
+]
+filterwarnings = [
+    "error",
+    "ignore::UserWarning",
+    "ignore::DeprecationWarning",
+    "ignore::PendingDeprecationWarning",
+]
+asyncio_mode = "auto"
+
+[tool.coverage.run]
+branch = true
+source = ["src"]
+omit = [
+    "*/tests/*",
+    "*/test_*",
+    "*/*_test.py",
+    "*/conftest.py",
+    "*/__pycache__/*",
+    "*/site-packages/*",
+    "*/.venv/*",
+    "*/venv/*",
+    "*/build/*",
+    "*/dist/*",
+    "*/.eggs/*",
+]
+
+[tool.coverage.report]
+skip_empty = true
+skip_covered = false
+show_missing = true
+sort = "Cover"
+exclude_lines = [
+    "pragma: no cover",
+    "def __repr__",
+    "if self.debug:",
+    "if settings.DEBUG",
+    "raise AssertionError",
+    "raise NotImplementedError",
+    "if 0:",
+    "if __name__ == .__main__.:",
+    "class .*\\bProtocol\\):",
+    "@(abc\\.)?abstractmethod",
+    "TYPE_CHECKING",
+]
+
+[tool.bandit]
+targets = ["src"]
+exclude_dirs = ["tests", "build", "dist"]
+skips = ["B101", "B601"]
+
+[tool.bandit.assert_used]
+skips = ["*/tests/*", "*/test_*.py"]
